Added function to clone a repository recursively
It is similar to `git clone --recursive`. The current implementation is very
simple and limited as it does not allow to customize the fetch and checkout
tasks and submodules are processed sequentially (i.e. no option like `--jobs`
to process them concurrently).

Addressing the limitations described above requires some thought:

- `FetchOptions` and `CheckoutBuilder` are not `Send`.
- Should different submodules have different options?
- Are new options and callbacks needed?